remove upgrade code for versions before 2.1
